+++ oe_patterns_cdk_common/asg.py
@@ -1,0 +1,9 @@
+from aws_cdk import (
+    aws_ec2,
+    core
+)
+
+class Asg(core.Construct):
+
+    def __init__(self, scope: core.Construct, id: str, **props):
+        super().__init__(scope, id)